backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
anthropic==0.39.0
pdfplumber==0.11.4
openpyxl==3.1.5
python-docx==1.1.2
pandas==2.2.3
python-dotenv==1.0.1
